Refactor: 241111 이미지 최적화 및 로직 수정 - image 압축 및 포맷타입 webp로 수정 - carousel 이미지 priority 옵션 추가하여 미리 로드되도록 구현 - Profile 컴포넌트 공통화 작업 - useFetchLikes : likes 조회 쿼리 생성 - useDoLike, useUndoLike : likes 이후 invalidate 쿼리키 'likes'로 수정 - feed/api : removeFeed 피드 삭제 시, 이미지 삭제되도록 로직 수정

+++ src/app/api/feed/update/route.ts
@@ -9,10 +9,10 @@ export async function PATCH(request: NextRequest) {
     const feedId = data.feedId;
     const delPaths = data.delPaths;
     const paths = data.paths;
-    console.log('data', data);
+    console.log('data', data, delPaths.length, paths.length);
     // 이미지 삭제
-    if (delPaths.length > 0) {
-      Promise.all(
+    if (delPaths && delPaths.length > 0) {
+      await Promise.all(
         delPaths.forEach(async (path: string) => {
           await db.query(
             `DELETE FROM images WHERE image_url = ? and feed_id = ?`,
@@ -25,8 +25,8 @@ export async function PATCH(request: NextRequest) {
     }
     console.log('[update img] delPaths=>', delPaths, 'paths=>', paths);
     // 이미지 추가
-    if (paths.length > 0) {
-      Promise.all(
+    if (paths && paths.length > 0) {
+      await Promise.all(
         paths.forEach(async (url: string) => {
           await db.query(
             `INSERT INTO images (feed_id, image_url) VALUES (?,?)`,
